# src/main/resources/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Web VolunteerDTO Coordination System - OpenAPI 3.0
  description: |-
    This is a Web Volunteer Coordination System based on the OpenAPI 3.0 specification. 
  contact:
    email: 260292@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: localhost:8080/
...
paths:
...
  /user:
    get:
      tags:
        - user
      summary: Get all users
      operationId: getUsers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags:
        - user
      summary: Create user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
  /user/{id}:
    put:
      tags:
        - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: name that need to be deleted
          required: true
          schema:
            type: integer
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '404':
          description: UserDTO not found
        '500':
          description: Internal server error
    delete:
      tags:
        - user
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: The id of user that needs to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '404':
          description: UserDTO not found
        '500':
          description: Internal server error
  /user/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      requestBody:
        description: The email and password for login
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: jack@gmail.com
                password:
                  type: string
                  example: '12345'
#        - name: email
#          in: query
#          description: The email for login
#          required: false
#          schema:
#            type: string
#        - name: password
#          in: query
#          description: The password for login in clear text
#          required: false
#          schema:
#            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid username/password supplied
...
  /volunteer:
    get:
      tags:
        - volunteer
      summary: Retrieve all volunteers
      operationId: getAllVolunteers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VolunteerDTO'

    post:
      tags:
        - volunteer
      summary: Create a new volunteer
      operationId: createVolunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerDTO'
      responses:
        '201':
          description: VolunteerDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerDTO'
  /volunteer/{volunteerId}:
    get:
      tags:
        - volunteer
      summary: Retrieve detailed volunteer info by ID
      operationId: getVolunteerById
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of the volunteer to retrieve info about
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerWithUserDTO'

    put:
      tags:
        - volunteer
      summary: Update a volunteer
      operationId: updateVolunteer
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of the volunteer to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerDTO'
      responses:
        '200':
          description: VolunteerDTO updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerDTO'

    delete:
      tags:
        - volunteer
      summary: Delete a volunteer
      operationId: deleteVolunteer
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of the volunteer to delete
          schema:
            type: integer
      responses:
        '204':
          description: VolunteerDTO deleted successfully
        '404':
          description: VolunteerDTO not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    VolunteerDTO:
      type: object
      required:
        - name
        - surname
        - username
        - userId
      properties:
        id:
          type: integer
          format: int64
          example: 102856
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        username:
          type: string
          example: joey
        userId:
          type: integer
          format: int64
          example: 100320
      xml:
        name: volunteer
...
    UserDTO:
      type: object
      required:
        - name
        - email
        - password
        - phoneNumber
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phoneNumber:
          type: string
          example: '12345'
      xml:
        name: user
...
    VolunteerWithUserDTO:
      type: object
      properties:
        volunteer:
          $ref: '#/components/schemas/VolunteerDTO'
        user:
          $ref: '#/components/schemas/UserDTO'
...